feat(ui): move modal component to ui package
