+++ packages/renderer/src/lib/extensions/InstallManuallyExtensionModal.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 import { faCloudDownload } from '@fortawesome/free-solid-svg-icons';
-import { Button, Input } from '@podman-desktop/ui-svelte';
+import { Button, Input, Modal } from '@podman-desktop/ui-svelte';
 import { onMount } from 'svelte';
 
-import Modal from '/@/lib/dialogs/Modal.svelte';
 import CloseButton from '/@/lib/ui/CloseButton.svelte';
 
 export let closeCallback: () => void;
