+++ packages/renderer/src/lib/feedback/SendFeedback.svelte
@@ -1,12 +1,11 @@
 <script lang="ts">
 import { faFrown, faGrinStars, faMeh, faSmile } from '@fortawesome/free-solid-svg-icons';
-import { Button, ErrorMessage } from '@podman-desktop/ui-svelte';
+import { Button, ErrorMessage, Modal } from '@podman-desktop/ui-svelte';
 import Fa from 'svelte-fa';
 
 import CloseButton from '/@/lib/ui/CloseButton.svelte';
 
 import type { FeedbackProperties } from '../../../../preload/src/index';
-import Modal from '../dialogs/Modal.svelte';
 import WarningMessage from '../ui/WarningMessage.svelte';
 
 let displayModal = false;
