+++ packages/renderer/src/lib/image/PushImageModal.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 import { faCheckCircle, faCircleArrowUp, faTriangleExclamation } from '@fortawesome/free-solid-svg-icons';
-import { Button } from '@podman-desktop/ui-svelte';
+import { Button, Modal } from '@podman-desktop/ui-svelte';
 import { onMount, tick } from 'svelte';
 import Fa from 'svelte-fa';
 import { router } from 'tinro';
@@ -10,7 +10,6 @@ import { FitAddon } from 'xterm-addon-fit';
 import CloseButton from '/@/lib/ui/CloseButton.svelte';
 
 import { TerminalSettings } from '../../../../main/src/plugin/terminal-settings';
-import Modal from '../dialogs/Modal.svelte';
 import Link from '../ui/Link.svelte';
 import type { ImageInfoUI } from './ImageInfoUI';
