+++ packages/renderer/src/lib/troubleshooting/TroubleshootingPageStoreDetails.svelte
@@ -1,13 +1,11 @@
 <script lang="ts">
 import { faRefresh, faTrash } from '@fortawesome/free-solid-svg-icons';
-import { Button } from '@podman-desktop/ui-svelte';
+import { Button, Modal } from '@podman-desktop/ui-svelte';
 import humanizeDuration from 'humanize-duration';
 import moment from 'moment';
 
 import CloseButton from '/@/lib/ui/CloseButton.svelte';
 import type { EventStoreInfo } from '/@/stores/event-store';
-
-import Modal from '../dialogs/Modal.svelte';
 
 export let closeCallback: () => void;
